@@ -8,7 +8,7 @@ import 'package:flutter_deriv_api/api/common/trading/trading_times.dart';
 import 'package:intl/intl.dart';
 
 /// Markets status change callback. (List of symbols that have been changed.)
-typedef OnMarketsStatusChange = void Function(Map<String, bool> symbols);
+typedef OnMarketsStatusChange = void Function(Map<String?, bool>? symbols);
 
 /// A class to remind when there is a change on market.
 ///
@@ -42,24 +42,24 @@ class MarketChangeReminder {
   /// Callback to get server time
   ///
   /// If not set it will be using DateTime.now().toUTC();
-  final Future<DateTime> Function() onCurrentTime;
+  final Future<DateTime> Function()? onCurrentTime;
 
   /// Callback to get trading times of today
   final Future<TradingTimes> Function() onTradingTimes;
 
   // TODO(Ramin): Consider using a reliable timer if Dart's version had any problems.
-  Timer _reminderTimer;
+  Timer? _reminderTimer;
 
   /// Gets called when market status changes with the list of symbols that their
   /// open/closed status is changed.
-  final OnMarketsStatusChange onMarketsStatusChange;
+  final OnMarketsStatusChange? onMarketsStatusChange;
 
   /// List of upcoming market change times.
   ///
   /// [SplayTreeMap] has been used to have the change times in correct order
   /// while we are adding new entries to it.
-  final SplayTreeMap<DateTime, Map<String, bool>> statusChangeTimes =
-      SplayTreeMap<DateTime, Map<String, bool>>();
+  final SplayTreeMap<DateTime, Map<String?, bool>> statusChangeTimes =
+      SplayTreeMap<DateTime, Map<String?, bool>>();
 
   Future<void> _init() async {
     await _fillStatusChangeMap();
@@ -71,29 +71,29 @@ class MarketChangeReminder {
 
     final TradingTimes todayTradingTimes = await onTradingTimes();
 
-    for (final MarketModel market in todayTradingTimes.markets) {
-      for (final SubmarketModel subMarket in market.submarkets) {
-        for (final SymbolModel symbol in subMarket.symbols) {
-          final List<dynamic> openTimes = symbol.times.open;
-          final List<dynamic> closeTimes = symbol.times.close;
+    for (final MarketModel? market in todayTradingTimes.markets!) {
+      for (final SubmarketModel? subMarket in market!.submarkets!) {
+        for (final SymbolModel? symbol in subMarket!.symbols!) {
+          final List<dynamic> openTimes = symbol!.times!.open!;
+          final List<dynamic>? closeTimes = symbol.times!.close;
 
           final bool isOpenAllDay = openTimes.length == 1 &&
               openTimes[0] == '00:00:00' &&
-              closeTimes[0] == '23:59:59';
+              closeTimes![0] == '23:59:59';
           final bool isClosedAllDay = openTimes.length == 1 &&
-              closeTimes[0] == '--' &&
+              closeTimes![0] == '--' &&
               closeTimes[0] == '--';
 
           if (isOpenAllDay || isClosedAllDay) {
             continue;
           }
 
-          for (final String time in openTimes) {
-            _addEntryToStatusChanges(time, symbol.symbol, true, now);
+          for (final String? time in openTimes as Iterable<String?>) {
+            _addEntryToStatusChanges(time!, symbol.symbol, true, now);
           }
 
-          for (final String time in closeTimes) {
-            _addEntryToStatusChanges(time, symbol.symbol, false, now);
+          for (final String? time in closeTimes as Iterable<String?>) {
+            _addEntryToStatusChanges(time!, symbol.symbol, false, now);
           }
         }
       }
@@ -102,7 +102,7 @@ class MarketChangeReminder {
 
   void _addEntryToStatusChanges(
     String time,
-    String symbol,
+    String? symbol,
     bool goesOpen,
     DateTime now,
   ) {
@@ -124,9 +124,9 @@ class MarketChangeReminder {
       return;
     }
 
-    statusChangeTimes[statusChangeTime] ??= <String, bool>{};
+    statusChangeTimes[statusChangeTime] ??= <String?, bool>{};
 
-    statusChangeTimes[statusChangeTime][symbol] = goesOpen;
+    statusChangeTimes[statusChangeTime]![symbol] = goesOpen;
   }
 
   /// Removes the next upcoming market change time from [statusChangeTimes] and
@@ -137,8 +137,8 @@ class MarketChangeReminder {
     final DateTime now = await _getNowTime();
 
     if (statusChangeTimes.isNotEmpty) {
-      final DateTime nextStatusChangeTime = statusChangeTimes.firstKey();
-      final Map<String, bool> symbolsChanging =
+      final DateTime nextStatusChangeTime = statusChangeTimes.firstKey()!;
+      final Map<String?, bool>? symbolsChanging =
           statusChangeTimes.remove(nextStatusChangeTime);
 
       _reminderTimer = Timer(
@@ -158,7 +158,7 @@ class MarketChangeReminder {
   }
 
   Future<DateTime> _getNowTime() async =>
-      onCurrentTime != null ? await onCurrentTime() : DateTime.now().toUtc();
+      onCurrentTime != null ? await onCurrentTime!() : DateTime.now().toUtc();
 
   /// Cancels current reminder timer.
   void reset() => _reminderTimer?.cancel();
